[SPARK-44496][SQL][CONNECT] Move Interfaces needed by SCSC to sql/api.
