+++ sql/api/src/main/java/org/apache/spark/api/java/function/MapPartitionsFunction.java
