+++ sql/api/src/main/java/org/apache/spark/api/java/function/VoidFunction2.java
