+++ sql/api/src/main/scala/org/apache/spark/sql/SqlApiConf.scala
@@ -40,6 +40,7 @@ private[sql] trait SqlApiConf {
   def doubleQuotedIdentifiers: Boolean
   def timestampType: AtomicType
   def allowNegativeScaleOfDecimalEnabled: Boolean
+  def charVarcharAsString: Boolean
 }
 
 private[sql] object SqlApiConf {
@@ -74,4 +75,5 @@ private[sql] object DefaultSqlApiConf extends SqlApiConf {
   override def doubleQuotedIdentifiers: Boolean = false
   override def timestampType: AtomicType = TimestampType
   override def allowNegativeScaleOfDecimalEnabled: Boolean = false
+  override def charVarcharAsString: Boolean = false
 }
