+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/expressions/ordering.scala
@@ -91,14 +91,7 @@ object RowOrdering extends CodeGeneratorWithInterpretedFallback[Seq[SortOrder], 
   /**
    * Returns true iff the data type can be ordered (i.e. can be sorted).
    */
-  def isOrderable(dataType: DataType): Boolean = dataType match {
-    case NullType => true
-    case dt: AtomicType => true
-    case struct: StructType => struct.fields.forall(f => isOrderable(f.dataType))
-    case array: ArrayType => isOrderable(array.elementType)
-    case udt: UserDefinedType[_] => isOrderable(udt.sqlType)
-    case _ => false
-  }
+  def isOrderable(dataType: DataType): Boolean = OrderUtils.isOrderable(dataType)
 
   /**
    * Returns true iff outputs from the expressions can be ordered.
